// src/app/employerPart/talent-board/model/canditate.model.ts
export interface Candidate {
    id: number;
    name: string;
    avatar: string;
    title: string;
    employer: string;
    location: string;
    yearsOfExperience: number;
    occupation: string; 
    education: string;
    skills: string[];
    availability: string;
    projectCount: number;
    rating: number;
  }